requirements.txt
attrs==21.2.0
certifi==2021.5.30
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
urllib3==1.26.3
pandas==1.5.2
numpy>=1.20.3
matplotlib==3.4.2
seaborn==0.11.1
xgboost==1.4.2
scikit-learn>=0.24.2
imbalanced-learn==0.8.0
scipy==1.6.3
imblearn==0.0
